refactor: split ResourceSystem.ts into smaller modules
- Extract ComputingManager for compute resource operations (131 lines)
- Extract DataManager for data resource management (254 lines)
- Extract ResourceCalculations for metrics and growth calculations (123 lines)
- Extract ResourceEffects for deployment/research effects (143 lines)
- Extract ResourceOperations for spending/affordability (138 lines)
- Main ResourceSystem now coordinates modules (187 lines, was 802)

This improves maintainability by separating concerns and making each module focused on a specific aspect of resource management.
