+++ src/systems/resources/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Barrel export for resource system modules
+ */
+
+export { ResourceCalculations } from './ResourceCalculations';
+export { ComputingManager } from './ComputingManager';
+export { DataManager } from './DataManager';
+export { ResourceEffects, ResourceEffectsManager } from './ResourceEffects';
+export { ResourceOperations } from './ResourceOperations';
